requirements.txt
streamlit>=1.31.0
google-genai>=0.1.0
openai>=1.12.0
apscheduler>=3.10.4
duckduckgo-search>=4.4.3
python-dotenv>=1.0.1
